backend/llm-api/requirements.txt
# =============================================================================
# Sabhya AI v0.4.0 - Requirements (Python 3.13 Compatible)
# =============================================================================
# All versions updated for Python 3.13 pre-built wheel support on Fedora 43
# =============================================================================

# Core Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Data Validation - IMPORTANT: Python 3.13 requires pydantic 2.9+
pydantic>=2.9.0
pydantic-settings>=2.3.0

# Database
sqlalchemy>=2.0.25
psycopg[binary]>=3.1.14

# Logging
structlog>=24.1.0
python-dotenv>=1.0.1

# AI & RAG - Python 3.13 compatible versions
chromadb>=0.5.0
pypdf>=4.0.0
tiktoken>=0.7.0
httpx>=0.26.0

# Rate Limiting
slowapi>=0.1.9

# Compatibility fixes
posthog<3.0.0
prometheus_client
numpy>=2.0

# ===== v0.4.0 Security Hardening =====
# PII Detection (Presidio NLP)
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl
presidio-analyzer>=2.2.35
presidio-anonymizer>=2.2.35

# JWT Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Distributed Rate Limiting
redis>=5.0.1

# Database Migrations
alembic>=1.13.1

# Monitoring
psutil>=5.9.8

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.3


email-validator
argon2-cffi
